requirements.txt
torch
diffusers
Pillow